dnsq: reduce usage of select to once
select() seems to work only once for our UDP sockets, perhaps because
they are unconnected.  Either way, use select() to wait for the first
answer after a query, then read non-blocking any responses that followed
the first (quick enough) when the first happens to be bad.  After that,
reconnect, so we can use select() again to wait for the first response.